Build support for linux | arm
The ${CSDK_HOME}/incl/cli directory was not being added
to the include path and was causing the build to fail on
Raspbian Jessie on a Model 2 B Raspberry Pi.

A condition check for OS=linux and target_arch=arm needed
to be added in binding.gyp. The build works as expected now.